+++ src/Symfony/Component/Form/ChoiceList/Factory/DefaultChoiceListFactory.php
@@ -52,7 +52,7 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
         return new LazyChoiceList($loader, $value);
     }
 
-    public function createView(ChoiceListInterface $list, array|callable|null $preferredChoices = null, callable|false|null $label = null, ?callable $index = null, ?callable $groupBy = null, array|callable|null $attr = null, array|callable $labelTranslationParameters = [], bool $duplicatePreferredChoices = true): ChoiceListView
+    public function createView(ChoiceListInterface $list, array|callable|null $preferredChoices = null, callable|false|null $label = null, ?callable $index = null, ?callable $groupBy = null, array|callable|null $attr = null, array|callable $labelTranslationParameters = [], bool $duplicatePreferredChoices = true, array|callable|null $help = null): ChoiceListView
     {
         $preferredViews = [];
         $preferredViewsOrder = [];
@@ -94,6 +94,7 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
                     $preferredViewsOrder,
                     $otherViews,
                     $duplicatePreferredChoices,
+                    $help
                 );
             }
 
@@ -133,6 +134,7 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
                 $preferredViewsOrder,
                 $otherViews,
                 $duplicatePreferredChoices,
+                $help
             );
         }
 
@@ -141,7 +143,7 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
         return new ChoiceListView($otherViews, $preferredViews);
     }
 
-    private static function addChoiceView($choice, string $value, $label, array $keys, &$index, $attr, $labelTranslationParameters, ?callable $isPreferred, array &$preferredViews, array &$preferredViewsOrder, array &$otherViews, bool $duplicatePreferredChoices): void
+    private static function addChoiceView($choice, string $value, $label, array $keys, &$index, $attr, $labelTranslationParameters, ?callable $isPreferred, array &$preferredViews, array &$preferredViewsOrder, array &$otherViews, bool $duplicatePreferredChoices, mixed $help): void
     {
         // $value may be an integer or a string, since it's stored in the array
         // keys. We want to guarantee it's a string though.
@@ -175,7 +177,8 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
             \is_callable($attr) ? $attr($choice, $key, $value) : ($attr[$key] ?? []),
             // The label translation parameters may be a callable or a mapping from choice indices
             // to nested arrays
-            \is_callable($labelTranslationParameters) ? $labelTranslationParameters($choice, $key, $value) : ($labelTranslationParameters[$key] ?? [])
+            \is_callable($labelTranslationParameters) ? $labelTranslationParameters($choice, $key, $value) : ($labelTranslationParameters[$key] ?? []),
+            \is_callable($help) ? $help($choice, $key, $value) : ($help[$key] ?? null),
         );
 
         // $isPreferred may be null if no choices are preferred
@@ -191,7 +194,7 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
         }
     }
 
-    private static function addChoiceViewsFromStructuredValues(array $values, $label, array $choices, array $keys, &$index, $attr, $labelTranslationParameters, ?callable $isPreferred, array &$preferredViews, array &$preferredViewsOrder, array &$otherViews, bool $duplicatePreferredChoices): void
+    private static function addChoiceViewsFromStructuredValues(array $values, $label, array $choices, array $keys, &$index, $attr, $labelTranslationParameters, ?callable $isPreferred, array &$preferredViews, array &$preferredViewsOrder, array &$otherViews, bool $duplicatePreferredChoices, mixed $help): void
     {
         foreach ($values as $key => $value) {
             if (null === $value) {
@@ -216,6 +219,7 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
                     $preferredViewsOrder,
                     $otherViewsForGroup,
                     $duplicatePreferredChoices,
+                    $help
                 );
 
                 if (\count($preferredViewsForGroup) > 0) {
@@ -243,11 +247,12 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
                 $preferredViewsOrder,
                 $otherViews,
                 $duplicatePreferredChoices,
+                $help,
             );
         }
     }
 
-    private static function addChoiceViewsGroupedByCallable(callable $groupBy, $choice, string $value, $label, array $keys, &$index, $attr, $labelTranslationParameters, ?callable $isPreferred, array &$preferredViews, array &$preferredViewsOrder, array &$otherViews, bool $duplicatePreferredChoices): void
+    private static function addChoiceViewsGroupedByCallable(callable $groupBy, $choice, string $value, $label, array $keys, &$index, $attr, $labelTranslationParameters, ?callable $isPreferred, array &$preferredViews, array &$preferredViewsOrder, array &$otherViews, bool $duplicatePreferredChoices, mixed $help): void
     {
         $groupLabels = $groupBy($choice, $keys[$value], $value);
 
@@ -266,6 +271,7 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
                 $preferredViewsOrder,
                 $otherViews,
                 $duplicatePreferredChoices,
+                $help,
             );
 
             return;
@@ -297,6 +303,7 @@ class DefaultChoiceListFactory implements ChoiceListFactoryInterface
                 $preferredViewsOrder[$groupLabel],
                 $otherViews[$groupLabel]->choices,
                 $duplicatePreferredChoices,
+                $help,
             );
         }
     }